nu-cli: merge completions tests into one file
