--- crates/nu-cli/tests/flag_completions.rs
@@ -1,38 +1,0 @@
-pub mod support;
-
-use nu_cli::NuCompleter;
-use reedline::Completer;
-use support::{match_suggestions, new_engine};
-
-#[test]
-fn flag_completions() {
-    // Create a new engine
-    let (_, _, engine, stack) = new_engine();
-
-    // Instatiate a new completer
-    let mut completer = NuCompleter::new(std::sync::Arc::new(engine), stack);
-    // Test completions for the 'ls' flags
-    let suggestions = completer.complete("ls -", 4);
-
-    assert_eq!(14, suggestions.len());
-
-    let expected: Vec<String> = vec![
-        "--all".into(),
-        "--directory".into(),
-        "--du".into(),
-        "--full-paths".into(),
-        "--help".into(),
-        "--long".into(),
-        "--short-names".into(),
-        "-D".into(),
-        "-a".into(),
-        "-d".into(),
-        "-f".into(),
-        "-h".into(),
-        "-l".into(),
-        "-s".into(),
-    ];
-
-    // Match results
-    match_suggestions(expected, suggestions);
-}
